[refactor] 약속 참여 페이지 리팩토링
